+++ apps/cowswap-frontend/src/modules/hooksStore/containers/PostHookButton/index.tsx
@@ -3,10 +3,11 @@ import { useWalletInfo } from '@cowprotocol/wallet'
 
 import SVG from 'react-inlinesvg'
 
+import { POST_HOOK_REGISTRY } from '../../hookRegistry'
 import { useHooks } from '../../hooks/useHooks'
 import { useRemoveHook } from '../../hooks/useRemoveHook'
 import { useReorderHooks } from '../../hooks/useReorderHooks'
-import { AppliedHookItem } from '../../pure/AppliedHookItem'
+import { AppliedHookList } from '../../pure/AppliedHookList'
 import { HookTooltip } from '../../pure/HookTooltip'
 import * as styledEl from '../PreHookButton/styled'
 
@@ -15,29 +16,27 @@ export interface PostHookButtonProps {
   onEditHook(uuid: string): void
 }
 
+const isPreHook = false
+
 export function PostHookButton({ onOpen, onEditHook }: PostHookButtonProps) {
-  const { account } = useWalletInfo()
+  const { account, chainId } = useWalletInfo()
   const { postHooks } = useHooks()
-  const removeHook = useRemoveHook()
+  const removeHook = useRemoveHook(isPreHook)
   const moveHook = useReorderHooks('postHooks')
+  const dapps = POST_HOOK_REGISTRY[chainId]
 
   return (
     <>
-      {postHooks && (
-        <styledEl.HookList>
-          {postHooks.map((hook, index) => (
-            <AppliedHookItem
-              key={hook.uuid}
-              account={account}
-              hookDetails={hook}
-              removeHook={removeHook}
-              editHook={onEditHook}
-              isPreHook={false}
-              index={index}
-              moveHook={moveHook}
-            />
-          ))}
-        </styledEl.HookList>
+      {postHooks.length > 0 && (
+        <AppliedHookList
+          dapps={dapps}
+          account={account}
+          hooks={postHooks}
+          isPreHook={false} // Indicate that these are post-hooks
+          removeHook={removeHook}
+          editHook={onEditHook}
+          moveHook={moveHook}
+        />
       )}
       <styledEl.Wrapper>
         <styledEl.AddHookButton onClick={onOpen}>