+++ apps/cowswap-frontend/src/modules/hooksStore/dapps/BuildHookApp/index.tsx
@@ -12,7 +12,7 @@ const DEFAULT_HOOK_STATE = {
   gasLimit: '',
 }
 
-export function BuildHookApp({ isPreHook, context }: HookDappProps) {
+export function BuildHookApp({ context }: HookDappProps) {
   const hookToEdit = context.hookToEdit
   const [hook, setHook] = useState<CowHook>(hookToEdit?.hook || DEFAULT_HOOK_STATE)
 
@@ -24,8 +24,11 @@ export function BuildHookApp({ isPreHook, context }: HookDappProps) {
       return
     }
 
-    context.editHook(hookToEdit.uuid, hook, isPreHook)
-  }, [hook, context, hookToEdit, isPreHook])
+    context.editHook({
+      ...hookToEdit,
+      hook,
+    })
+  }, [hook, context, hookToEdit])
 
   const clickOnAddHook = useCallback(() => {
     const { callData, gasLimit, target } = hook
@@ -74,7 +77,7 @@ export function BuildHookApp({ isPreHook, context }: HookDappProps) {
       {hookToEdit ? (
         <ButtonPrimary onClick={onEditHook}>Save changes</ButtonPrimary>
       ) : (
-        <ButtonPrimary onClick={clickOnAddHook}>Add {isPreHook ? 'Pre' : 'Post'}-hook</ButtonPrimary>
+        <ButtonPrimary onClick={clickOnAddHook}>Add {context.isPreHook ? 'Pre' : 'Post'}-hook</ButtonPrimary>
       )}
     </Wrapper>
   )