+++ apps/cowswap-frontend/src/modules/hooksStore/pure/AppliedHookItem/index.tsx
@@ -6,40 +6,37 @@ import { Edit2, Trash2 } from 'react-feather'
 import SVG from 'react-inlinesvg'
 
 import * as styledEl from './styled'
-import { useDragAndDrop } from './useDragAndDrop'
 
 import { TenderlySimulate } from '../../containers/TenderlySimulate'
-import { CowHookDetailsSerialized } from '../../types/hooks'
+import { CowHookDetailsSerialized, HookDapp } from '../../types/hooks'
 
 interface HookItemProp {
   account: string | undefined
   hookDetails: CowHookDetailsSerialized
+  dapp: HookDapp
   isPreHook: boolean
   removeHook: (uuid: string, isPreHook: boolean) => void
   editHook: (uuid: string) => void
   index: number
-  moveHook: (dragIndex: number, hoverIndex: number) => void
 }
 
 export function AppliedHookItem({
   account,
-  hookDetails,
+  hookDetails: { hookDetails },
+  dapp,
   isPreHook,
   editHook,
   removeHook,
   index,
-  moveHook,
 }: HookItemProp) {
-  const { hook, dapp } = hookDetails
-  const { ref, isDragging } = useDragAndDrop(index, hookDetails.uuid, moveHook)
-
   return (
-    <styledEl.HookItemWrapper data-uid={hookDetails.uuid} ref={ref} style={{ opacity: isDragging ? 0.5 : 1 }}>
+    <styledEl.HookItemWrapper data-uid={hookDetails.uuid} as="li">
       <styledEl.HookItemHeader title={hookDetails.uuid}>
-        <styledEl.HookItemInfo>
+        <styledEl.HookItemInfo className="DragArea">
           <styledEl.DragIcon>
             <SVG src={ICON_GRID} />
           </styledEl.DragIcon>
+          <styledEl.HookNumber>{index + 1}</styledEl.HookNumber>
           <img src={dapp.image} alt={dapp.name} />
           <span>{dapp.name}</span>
         </styledEl.HookItemInfo>
@@ -66,7 +63,7 @@ export function AppliedHookItem({
             </styledEl.SimulateFooter>
           </div>
           <div>
-            <TenderlySimulate hook={hook} />
+            <TenderlySimulate hook={hookDetails.hook} />
           </div>
         </styledEl.SimulateContainer>
       )}