+++ libs/events/src/types/widgetEvents.ts
@@ -1,0 +1,6 @@
+import { CowWidgetEventPayloadMap, CowWidgetEvents } from './events'
+
+import { CowEventListener } from '../CowEventEmitter'
+
+export type CowWidgetEventListener = CowEventListener<CowWidgetEventPayloadMap, CowWidgetEvents>
+export type CowWidgetEventListeners = CowWidgetEventListener[]