+++ libs/events/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.4.1](https://github.com/cowprotocol/cowswap/compare/events-v1.4.0...events-v1.4.1) (2024-09-17)
+
+
+### Bug Fixes
+
+* **libs:** fix import error on multiple libs ([#4869](https://github.com/cowprotocol/cowswap/issues/4869)) ([8cfd70b](https://github.com/cowprotocol/cowswap/commit/8cfd70b6000a1a021b76ad1f924829cb42cb2bcd))
+
 ## [1.4.0](https://github.com/cowprotocol/cowswap/compare/events-v1.3.0...events-v1.4.0) (2024-07-12)